Add nuget publishing to release action (#3)
* chore: Add workflow for publishing NuGet packages

The changes in the `dotnet-release.yml` file introduce a new workflow for publishing NuGet packages. Specifically, the following modifications were made:
- Added a trigger for the workflow to run on the creation of any tags.
- Introduced a new job named `publish_nuget` that depends on the `semantic_release` job.
- The `publish_nuget` job runs on the `ubuntu-latest` environment.
- Steps for the `publish_nuget` job include:
  - Checking out the repository using `actions/checkout@v2`.
  - Setting up the .NET environment with version `8.x` using `actions/setup-dotnet@v1`.
  - Building the NuGet package with the `dotnet pack` command, setting the package version to the latest tag.
  - Publishing the NuGet package to the NuGet repository using the `dotnet nuget push` command, with the API key stored in the `NUGET_API_KEY` secret.

* chore: Use latest Git tag as version in dotnet pack

Added a new step to retrieve the latest Git tag and set it as an output variable. Updated the `dotnet pack` command to use this latest Git tag as the package version.

* chore: Simplify latest Git tag retrieval in dotnet-release.yml

Updated the method for retrieving the latest Git tag in the
dotnet-release.yml file. Replaced the multi-line script with
a single-line command using `git describe --tags` to set the
latest_tag output. Corrected the position of the `id` for the
`Get latest tag` step.

* chore: Update GitHub Actions workflow triggers

The `dotnet-release.yml` file has been updated to modify the GitHub Actions workflow configuration. The `push` section now includes branches `master`, `beta`, and `pre-release`. Additionally, a `create` section has been added to include `tags` with a wildcard pattern `'*'`. These changes ensure that the build process is initiated for the specified branches and any new tags.