chore: Simplify tag handling and conditional publishing
Simplified the `echo` command for `LATEST_TAG_BEFORE_SEMANTIC_RELEASE` by removing unnecessary double quotes. Updated the `Check if tags are different` step to set an output `SKIP_PUBLISH` to `true` instead of exiting with an error code when tags are the same or when `LATEST_TAG_BEFORE_SEMANTIC_RELEASE` is empty. Added conditional checks to the `Build NuGet package` and `Publish NuGet package` steps to only run if `SKIP_PUBLISH` is not `true`.