chore(release): Add step to echo head branch name in GitHub Actions
Added a step to print the head branch name in the GitHub Actions
workflow using the `echo` command. Adjusted the indentation for the
`ref` parameter in the `actions/checkout` step to align with the new
step. The `ref` parameter continues to use the `${{ github.event.workflow_run.head_branch }}`
expression to determine the branch to check out.